src/requirements.txt
aiohttp~=3.9.4
python-dotenv
chardet